requirements.txt
apache-airflow==2.9.1
pandas
snowflake-connector-python
snowflake-sqlalchemy

